requirements.txt
asgiref==3.5.2
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.12
cycler==0.11.0
Django==4.0.5
django-dotenv==1.4.2
django-filter==21.1
django-grappelli==2.15.1
django-htmx==1.11.0
django-multiselectfield==0.1.12
django-suit==2.0a1
djangorestframework==3.13.1
easyocr==1.5.0
fonttools==4.33.3
idna==3.3
imageio==2.19.3
kiwisolver==1.4.2
matplotlib==3.5.2
networkx==2.8.2
numpy==1.22.4
opencv-python==4.5.5.64
opencv-python-headless==4.5.4.60
packaging==21.3
pickleshare==0.7.5
Pillow==9.1.1
pkginfo==1.8.2
prompt-toolkit==3.0.28
psycopg2==2.9.3
Pygments==2.11.2
pyparsing==3.0.7
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.3.0
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.27.1
rfc3986==2.0.0
scikit-image==0.19.2
scipy==1.8.1
SimpleWebSocketServer==0.1.1
six==1.16.0
sqlparse==0.4.2
tifffile==2022.5.4
torch==1.11.0
torchvision==0.12.0
traitlets==5.1.1
typing==3.7.4.3
typing_extensions==4.1.1
tzdata==2022.1
urllib3==1.26.8
ushlex==0.99.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
whitenoise==6.0.0
zipp==3.7.0
